+++ changelog.md
@@ -10,13 +10,14 @@
   - Fixed: Containers and target objects are sometimes initially positioned in mid-air.
   - Fixed: Containers and target objects are sometimes initially positioned where the avatar can't reach them.
   - Reduced the list of possible target objects.
+  - `goal_positions` is now a dictionary of room indices, model names, and positions.
 
 ### Backend
 
 - Added: `precision_test.py` Test how precision affects arm articulation.
 - Added: `proc_gen_spawn_test.py` Test whether target objects and containers tend to stay in their initial positions.
-- Added: `reachable_surface_maps/` which includes arrays of booleans. If True, the position is a floor or low-lying surface.
-- `occupancy_mapper.py` generates more accurate maps and can create reachable surface maps (see above).
+- Added: `surface_maps/` which includes arrays of booleans. If True, the position is a floor or low-lying surface.
+- `occupancy_mapper.py` generates more accurate maps and creates surface maps.
 
 ## 0.6.2